% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_list}
\alias{benchmark_list}
\title{Benchmark List}
\format{
## `benchmark_list`
A character vector of length 18
}
\source{
Internal
}
\usage{
benchmark_list
}
\description{
A subset of data for benchmark
}
\keyword{datasets}
