% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{case_when_for_value_setting}
\alias{case_when_for_value_setting}
\title{Title}
\usage{
case_when_for_value_setting(dat, cn, col, new_col)
}
\arguments{
\item{dat}{dataframe}

\item{cn}{name of column as character}

\item{col}{column name as symbol}

\item{new_col}{new column name to be created passed as symbol}
}
\value{
Data frame
}
\description{
Title
}
