% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bm}
\alias{dat_bm}
\title{dat_bm}
\format{
## `dat_bm`
A dataframe with 7664 rows and 35 columns
}
\source{
Internal
}
\usage{
dat_bm
}
\description{
A subset of data for benchmark
}
\keyword{datasets}
