% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_country_income_region}
\alias{dat_country_income_region}
\title{dat_country_income_region}
\format{
## `dat_country_income_region`
A dataframe with 266 rows and 3 columns
}
\source{
Internal
}
\usage{
dat_country_income_region
}
\description{
Data for income region in country
}
\keyword{datasets}
