% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_layout}
\alias{dat_layout}
\title{dat_layout}
\format{
## `dat_layout`
List of 7 dataframes
}
\source{
Internal
}
\usage{
dat_layout
}
\description{
A list containing 7 dataframes
}
\keyword{datasets}
