% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ports}
\alias{dat_ports}
\title{dat_ports}
\format{
## `dat_ports`
A tibble with 12288 rows and 132 columns
}
\source{
Internal
}
\usage{
dat_ports
}
\description{
dat_ports
}
\keyword{datasets}
