% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ports_bm}
\alias{dat_ports_bm}
\title{dat_ports_bm}
\format{
## `dat_ports_bm`
A tibble with 218 rows and 43 columns
}
\source{
Internal
}
\usage{
dat_ports_bm
}
\description{
dat_ports_bm
}
\keyword{datasets}
