% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irf_indicators}
\alias{irf_indicators}
\title{irf_indicators}
\format{
## `irf_indicators`
A character vector of length 47
}
\source{
Internal
}
\usage{
irf_indicators
}
\description{
irf_indicators
}
\keyword{datasets}
