% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_location}
\alias{map_location}
\title{map_location}
\format{
## `map_location`
A tibble with 8 rows and 4 columns.
}
\source{
Internal
}
\usage{
map_location
}
\description{
map_location
}
\keyword{datasets}
