% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scd_dat_countries_with_article}
\alias{scd_dat_countries_with_article}
\title{scd_dat_countries_with_article}
\format{
## `scd_dat_countries_with_article`
A character vector of length 19.
}
\source{
Internal
}
\usage{
scd_dat_countries_with_article
}
\description{
scd_dat_countries_with_article
}
\keyword{datasets}
