% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scd_indicators}
\alias{scd_indicators}
\title{scd_indicators}
\format{
## `scd_indicators`
A tibble with 55 rows and 2 columns.
}
\source{
Internal
}
\usage{
scd_indicators
}
\description{
scd_indicators
}
\keyword{datasets}
